make it easier to add new launchers
