+++ launchers/launchers.go
@@ -1,53 +1,20 @@
 package launchers
 
-type Minecraft struct {
-	profile       string
-	version       string
-	loader        string
-	loader_version string
-	username      string
-	uuid          string
-	token         string
+import (
+	"errors"
+	"main/platform"
+	"os"
+)
+
+var LAUNCHERS = map[string]LauncherSupport{
+	PRISM_ID:    LauncherSupport{NewPrismLauncher, PRISM_NAME, PRISM_CLASS},
+	MODRINTH_ID: LauncherSupport{NewModrinthApp, MODRINTH_NAME, MODRINTH_CLASS},
 }
 
-func NewMinecraft(
-	instance Instance,
-) Minecraft {
-	return Minecraft{
-		instance.Name(),
-		instance.Version(),
-		instance.Loader(),
-		instance.LoaderVersion(),
-		"", "", "",
-	}
-}
-
-func (minecraft Minecraft) Profile() string {
-	return minecraft.profile
-}
-
-func (minecraft Minecraft) Version() string {
-	return minecraft.version
-}
-
-func (minecraft Minecraft) Loader() string {
-	return minecraft.loader
-}
-
-func (minecraft Minecraft) LoaderVersion() string {
-	return minecraft.loader_version
-}
-
-func (minecraft Minecraft) Username() string {
-	return minecraft.username
-}
-
-func (minecraft Minecraft) Uuid() string {
-	return minecraft.uuid
-}
-
-func (minecraft Minecraft) Token() string {
-	return minecraft.token
+type LauncherSupport struct {
+	New       func(string, string, bool) Launcher
+	Name      string
+	JavaClass string
 }
 
 type Launcher interface {
@@ -60,15 +27,16 @@ type Launcher interface {
 	GetInstances(map[string][]string, string) ([]Instance, error)
 	Exists() bool
 	IsRunning() bool
+	DefaultNotify() bool
 	MergeNormal() Launcher
 }
 
 type launcher_data struct {
-	name string
-	path string
+	name         string
+	path         string
 	instance_dir string
-	flatpak_id string
-	replaced bool
+	flatpak_id   string
+	replaced     bool
 }
 
 func (data launcher_data) Name() string {
@@ -95,4 +63,45 @@ func (data launcher_data) ReplaceNormal() {
 	if data.flatpak_id != "" {
 		data.replaced = true
 	}
+}
+
+func appendLaunchers(
+	launchers []Launcher,
+	ctor func(string, string, bool) Launcher,
+	home string,
+) []Launcher {
+	regular := ctor(home, "", false)
+	if platform.WINDOWS {
+		if regular.Exists() {
+			launchers = append(launchers, regular)
+		}
+	} else {
+		flatpak := ctor(home, "", true)
+		if regular.Exists() && flatpak.Exists() && regular.InstanceDir() == flatpak.InstanceDir() {
+			launchers = append(launchers, flatpak.MergeNormal())
+		} else {
+			if regular.Exists() {
+				launchers = append(launchers, regular)
+			}
+			if flatpak.Exists() {
+				launchers = append(launchers, flatpak)
+			}
+		}
+	}
+	return launchers
+}
+
+func GetLaunchers() ([]Launcher, error) {
+	home, err := os.UserHomeDir()
+	if err != nil {
+		return nil, err
+	}
+	var launchers []Launcher
+	for _, l := range LAUNCHERS {
+		launchers = appendLaunchers(launchers, l.New, home)
+	}
+	if len(launchers) == 0 {
+		return nil, errors.New("No launchers installed")
+	}
+	return launchers, nil
 }
